# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Booking API",
        "description": "API for managing hotel room bookings",
        "version": "1.0.0"
    },
    "paths": {
        "/api/bookings": {
            "get": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Get all active bookings",
                "operationId": "eb7c332e8ba161904a1c0071040f61a5",
                "responses": {
                    "200": {
                        "description": "List of bookings",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "guest_name": {
                                                        "type": "string"
                                                    },
                                                    "check_in_date": {
                                                        "type": "string",
                                                        "format": "date"
                                                    },
                                                    "check_out_date": {
                                                        "type": "string",
                                                        "format": "date"
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "enum": [
                                                            "pending",
                                                            "confirmed",
                                                            "cancelled"
                                                        ]
                                                    },
                                                    "promo_code": {
                                                        "type": "string",
                                                        "nullable": true
                                                    },
                                                    "room": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer"
                                                            },
                                                            "number": {
                                                                "type": "string"
                                                            },
                                                            "type": {
                                                                "type": "string"
                                                            },
                                                            "is_available": {
                                                                "type": "boolean"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Create a new booking",
                "operationId": "e33fa5a6507581f9b1e0d710a9a130d1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "guest_name",
                                    "room_id",
                                    "check_in_date",
                                    "check_out_date"
                                ],
                                "properties": {
                                    "guest_name": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "room_id": {
                                        "type": "integer"
                                    },
                                    "check_in_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "check_out_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "promo_code": {
                                        "type": "string",
                                        "maxLength": 50,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Booking created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Booking created"
                                        },
                                        "booking": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "guest_name": {
                                                    "type": "string"
                                                },
                                                "room_id": {
                                                    "type": "integer"
                                                },
                                                "check_in_date": {
                                                    "type": "string",
                                                    "format": "date"
                                                },
                                                "check_out_date": {
                                                    "type": "string",
                                                    "format": "date"
                                                },
                                                "status": {
                                                    "type": "string"
                                                },
                                                "promo_code": {
                                                    "type": "string",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict - Room not available"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Bookings",
            "description": "Bookings"
        }
    ]
}